changed col in matchedliveusers

+++ database/models/livecontestuserpool.js
@@ -56,7 +56,7 @@ export default (sequelize, DataTypes) => {
       userId: { type: DataTypes.STRING, allowNull: false },
       socketId: { type: DataTypes.STRING, allowNull: false },
       stockId: { type: DataTypes.INTEGER, allowNull: true },
-      stockValue: { type: DataTypes.INTEGER, allowNull: true },
+      stockValue: { type: DataTypes.FLOAT, allowNull: true },
       matched: { type: DataTypes.BOOLEAN, allowNull: true },
       isBot: { type: DataTypes.BOOLEAN, allowNull: true},
       contestEntryPrice: { type: DataTypes.FLOAT, allowNull:true},
